+++ plugins/nri/plugin.go
@@ -14,10 +14,10 @@
    limitations under the License.
 */
 
-package plugin
+package nri
 
 import (
-	"github.com/containerd/containerd/v2/pkg/nri"
+	"github.com/containerd/containerd/v2/internal/nri"
 	"github.com/containerd/containerd/v2/plugins"
 	"github.com/containerd/plugin"
 	"github.com/containerd/plugin/registry"
